refactor: 캔들 조회 어댑터 계층 개선 - FindCandlesRequest에서 도메인 변환 로직 및 기본값 제거 - CandleResponse 가격 필드 BigDecimal 전환 - InfluxDB range(start: 0) → 동적 range 계산으로 변경 - Flux filter 표현식 단일 블록으로 병합 - toDouble → toBigDecimal 변환, 유효하지 않은 값에 예외 발생

+++ src/main/java/ksh/tryptobackend/marketdata/adapter/in/dto/response/CandleResponse.java
@@ -2,14 +2,15 @@ package ksh.tryptobackend.marketdata.adapter.in.dto.response;
 
 import ksh.tryptobackend.marketdata.domain.model.Candle;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record CandleResponse(
     Instant time,
-    double open,
-    double high,
-    double low,
-    double close
+    BigDecimal open,
+    BigDecimal high,
+    BigDecimal low,
+    BigDecimal close
 ) {
 
     public static CandleResponse from(Candle candle) {
